api_for_mcp/save_to_file/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
